extracted display components from Feed to allow abstraction and leter redux
